Added prelimary servlet and java bean support
new servlets
- CategoryServlet: this handles all category page requests
- CheckoutServlet: unimplemented.. this will handle the checkout request
- LoginServlet: currently a shell - will handle login requests
- RegisterServlet: currently a shell - will handle registration requests

Beans added:
- CategoryWorkerBean: handles some of the processing between the
database and the creation of ItemBeans
- ItemBean: will be a java bean class containing information about items

TODO:
- create UserBean, DetailedItemBean (which will be a more complex query,
getting all information from all tables about the item), etc
- finish implementing the current servlets, add new ones as necessary

FOR JASON
- feel free to edit the ItemBean class to match your database design.
We'll need a field (with getters/setters) for each column in the Item
table.

FOR JIE
- I couldn't get the links to the servlet to work without adding a
"submit" button to each of the forms in the drop down menus. The current
buttons aren't very pretty, if you want to try to find a better way to
call the servlet or just make the button less hideous, that would be a
huge help.